fix model long, lat parsing issue

+++ Restaurant Project/Restaurant Project/RestaurantModel.swift
@@ -59,8 +59,8 @@ struct ContactInfo : Decodable {
 }
 
 struct LocationInfo : Decodable {
-    let lat : Int?
-    let lng : Int?
+    let lat : Double?
+    let lng : Double?
     let address : String?
     let city : String?
     let state : String?
